+++ AINDY/db/models/__init__.py
@@ -13,6 +13,7 @@ from .leadgen_model import LeadGenResult
 from .author_model import AuthorDB
 from .system_health_log import SystemHealthLog
 from .user import User
+from .user_identity import UserIdentity
 from .memory_node_history import MemoryNodeHistory
 
 
@@ -35,5 +36,6 @@ __all__ = [
     "SystemHealthLog",
     "GenesisSessionDB",
     "User",
+    "UserIdentity",
     "MemoryNodeHistory",
 ]